requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
python-dotenv==1.0.1
psycopg[binary]==3.2.3
gunicorn==22.0.0
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3
google-cloud-storage==2.18.2
google-cloud-tasks==2.17.0
pydantic==2.9.2
